--- src/CustomHooks/useWindowPosition.js
@@ -1,17 +1,0 @@
-import { useLayoutEffect, useState } from "react";
-
-export default function useWindowPosition(id) {
-    const [animation, setAnimation] = useState(false);
-
-    useLayoutEffect(() => {
-        function updatePosition() {
-            const offsetHeight = window.document.getElementById(id).offsetHeight;
-            if(window.pageYOffset > offsetHeight * 0.5)
-                setAnimation(true);
-        }
-        window.addEventListener('scroll', updatePosition);
-        updatePosition();
-        return () => window.removeEventListener('scroll', updatePosition);
-    }, [id])
-    return animation;
-}
