+++ src/Services/UserServices/UserAPI.js
@@ -1,17 +1,6 @@
 import { getCustomHttp } from "../../CustomHooks/useCustomHttp";
 class UserAPI {
-  addUser(users) {
-    const http = getCustomHttp();
-    http.post("/User/Adding/", {
-      Id: users.Id,
-      Name: users.Name,
-      Surname: users.Surname,
-      Birthdate: users.Birthdate,
-      Email: users.Email,
-      PhoneNumber: users.PhoneNumber,
-    });
-  } 
-  
+
   UpdateDriver(LicenseNumber, DriversLicense, VehicleType) {
     const http = getCustomHttp();
     http.put("/User/Driver", {
@@ -24,11 +13,6 @@ class UserAPI {
   GetUser() {
     const http = getCustomHttp();
     return http.get(`/User`);
-  }
-
-  GetAllUsers() {
-    const http = getCustomHttp();
-    return http.get("/User/Users");
   }
 }
