--- src/UIcomponents/GoogleMapIntegration/Styles/GoogleMapsStyle.module.css
@@ -1,45 +1,0 @@
-.map_container {
-    width: 100%;
-    height: 100vh;
-}
-
-.split_screen {
-    display: flex;
-    flex-direction: row;
-}
-
-.top_pane{
-    width: 50%;
-}
-
-.bottom_pane {
-    width: 50%;
-}
-
-.map_width {
-    width: 100%;
-}
-
-.address_field {
-    width: 98%;
-    margin-top: 3px;
-    margin-left: 10px;
-    border-radius: 30px;
-    padding-left: 10px
-}
-
-.createButton {
-    color: black !important;
-    border: 1px solid black !important;
-    background-color: #F0F0F0 !important;
-}
-
-.createButton:hover {
-    cursor: pointer;
-    margin-top: 3px;
-    border: 2px solid black;
-}
-
-.createButton:active {
-    background-color:beige;
-}
