--- src/UIcomponents/Styles/BackgroundStyle.module.css
@@ -1,7 +1,0 @@
-.bg {
-  min-height: 100vh;
-  background-image: url("../../Images/landing_background.jpg");
-  background-repeat: repeat;
-  background-size: cover;
-  height: 100%;
-}
