+++ src/UIcomponents/TripPlanning/PlanningComponent.jsx
@@ -1,4 +1,4 @@
-import { React, useMemo, useState, useEffect } from "react";
+import { React, useState, useEffect } from "react";
 import HeaderComponent from "../BaseHeader/HeaderComponent";
 import FooterComponent from "../BaseFooter/FooterComponent";
 import { useUserValidation } from "../../CustomHooks/useUserValidation";
